Moved ArticleDTO to API level
