+++ SportsHub/SportsHub.Api/Controllers/ArticleController.cs
@@ -1,6 +1,6 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
-using SportsHub.AppService.Authentication.Models.DTOs;
+using SportsHub.Api.Mapping.Models;
 using SportsHub.AppService.Services;
 
 namespace SportsHub.Api.Controllers
